requirements.txt
flask==3.0.0
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
flask-login==0.6.3
PyYAML>=6.0.1
scipy>=1.11.0 